requirements.txt
pytest
selenium
requests
mysql-connector-python
json
random